Do not intent to VectorMessagingReceiver from FirebaseReceiver
Use VectorMessagingHelper to directly call onMessage
